Documents/bitspilani/qpm/requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-bootstrap3==15.0.0
django-import-export==2.5.0
et-xmlfile==1.0.1
importlib-metadata==2.1.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
django-datatable
django-extensions
celery==4.4.7
django-celery-results==1.2.1
redis
python-magic==0.4.22
boto3==1.18.6
django-storages==1.11.1
djangorestframework
requests
